fix: split TablePanelForm action row to restore ESLint max-lines budget
Extract ActionRow into its own module so TablePanelFormRows stays under the 500-line limit.
